refactor: update to latest java sdk
