+++ workflow/java/pom.xml
@@ -42,7 +42,7 @@
     <dependency>
       <groupId>io.dapr</groupId>
       <artifactId>dapr-sdk-workflows</artifactId>
-      <version>0.11.0</version>
+      <version>0.14.0</version>
     </dependency>
   </dependencies>
