+++ workflow/java/src/main/java/io/dapr/quickstarts/workflows/WorkflowConfig.java
@@ -2,7 +2,7 @@ package io.dapr.quickstarts.workflows;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import io.dapr.workflows.runtime.WorkflowRuntime;
+import io.dapr.workflows.client.DaprWorkflowClient;
 import io.dapr.workflows.runtime.WorkflowRuntimeBuilder;
 
 import io.dapr.quickstarts.workflows.activities.*;
@@ -11,15 +11,18 @@ import io.dapr.quickstarts.workflows.activities.*;
 public class WorkflowConfig {
 
   @Bean
-  public WorkflowRuntime workflowRuntime() {
-    WorkflowRuntimeBuilder builder = new WorkflowRuntimeBuilder().registerWorkflow(OrderProcessingWorkflow.class);
+  public DaprWorkflowClient daprWorkflowClient() {
+    return new DaprWorkflowClient();
+  }
+
+  @Bean
+  public WorkflowRuntimeBuilder workflowRuntimeBuilder() {
+    WorkflowRuntimeBuilder builder = new WorkflowRuntimeBuilder();
+    builder.registerWorkflow(OrderProcessingWorkflow.class);
     builder.registerActivity(NotifyActivity.class);
     builder.registerActivity(ProcessPaymentActivity.class);
     builder.registerActivity(ReserveInventoryActivity.class);
     builder.registerActivity(UpdateInventoryActivity.class);
-
-    WorkflowRuntime runtime = builder.build();
-    runtime.start(false);
-    return runtime;
+    return builder;
   }
 }
