+++ workflow/java/src/main/java/io/dapr/quickstarts/workflows/activities/NotifyActivity.java
@@ -1,8 +1,8 @@
 package io.dapr.quickstarts.workflows.activities;
 
 import io.dapr.quickstarts.workflows.models.Notification;
-import io.dapr.workflows.runtime.WorkflowActivity;
-import io.dapr.workflows.runtime.WorkflowActivityContext;
+import io.dapr.workflows.WorkflowActivity;
+import io.dapr.workflows.WorkflowActivityContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
