+++ workflow/java/src/main/java/io/dapr/quickstarts/workflows/activities/UpdateInventoryActivity.java
@@ -1,7 +1,7 @@
 package io.dapr.quickstarts.workflows.activities;
 
-import io.dapr.workflows.runtime.WorkflowActivity;
-import io.dapr.workflows.runtime.WorkflowActivityContext;
+import io.dapr.workflows.WorkflowActivity;
+import io.dapr.workflows.WorkflowActivityContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
